requirements.txt
# Core
numpy==1.24.4
pillow==10.2.0
opencv-python==4.9.0.80

# PyTorch (CPU)
torch==1.13.1
torchvision==0.14.1

# YOLO
ultralytics==8.3.245

# Web App
streamlit==1.31.1
